Add persistence

+++ machi/machi.py
@@ -20,6 +20,7 @@ Data files has filename <gen>.machd
 
 '''
 import binascii
+import glob
 import os
 from struct import pack, unpack, calcsize
 from typing import Optional
@@ -28,38 +29,66 @@ from .rwlock import RWLock
 
 class _MachiGen:
     index_format = 'QQQIi'
-    
-    def __init__(self, dir, gen):
+    index_format_size = calcsize(index_format)
+
+    def __init__(self, dir, gen, creat=True, temp=False):
         self.gen = gen
         self.dir = dir
+        self.temp = temp
         self.index = {}
-        flag = os.O_RDWR|os.O_CREAT|os.O_EXCL|os.O_TRUNC
         self.indexname = os.path.join(self.dir, f'{self.gen}.machi')
-        self.indexfile = os.open(self.indexname, flag)
         self.dataname = os.path.join(self.dir, f'{self.gen}.machd')
-        self.datafile = os.open(self.dataname, flag)
-        self.size = 0
-        self._ref = 0
+        if creat:
+            flag = os.O_RDWR|os.O_CREAT|os.O_EXCL|os.O_TRUNC
+            self.indexfile = os.open(self.indexname, flag)
+            self.datafile = os.open(self.dataname, flag)
+            self._ref = 0
 
-        self.index_pos = 0
-        self.pos = 0
-        
+            self.index_pos = 0
+            self.pos = 0
+        else:
+            index_stat = os.stat(self.indexname)
+            self.index_pos = index_stat.st_size
+            if self.index_pos % self.index_format_size != 0:
+                print("Partial write happened previous era")
+                remain = self.index_pos % self.index_format_size
+                self.index_pos -= remain
+
+            data_stat = os.stat(self.dataname)
+            self.pos = data_stat.st_size
+
+            flag = os.O_RDWR|os.O_EXCL
+            self.indexfile = os.open(self.indexname, flag)
+
+            index_pos = 0
+            while index_pos < self.index_pos:
+                data = os.pread(self.indexfile,
+                                self.index_format_size,
+                                index_pos)
+                g, o, l, c, s = unpack(self.index_format, data)
+                assert g == gen
+                self.index[o] = l, c, s, index_pos
+
+                index_pos += self.index_format_size
+
+            self.datafile = os.open(self.dataname, flag)
+            self._ref = 0
+
     def __len__(self):
-        return self.size
-    
+        return len(self.index) - self._ref
+
     @property
     def ref(self):
         return self._ref
 
     def append(self, data):
-        self.size += 1
         self._ref += 1
-        
+
         pos = self.pos
         r = os.pwrite(self.datafile, data, pos)
         assert len(data) == r
         self.pos += len(data)
-        
+
         crc = binascii.crc32(data)
         buf = pack(self.index_format, self.gen, pos,
                    len(data), crc, 1)
@@ -70,11 +99,17 @@ class _MachiGen:
         self.index_pos += r
 
         return pos, len(data)
-    
+
+    def keys(self):
+        for pos in self.index.keys():
+            l, c, s, p = self.index[pos]
+            if s == 1:
+                yield self.gen, pos, l
+
     def get(self, offset, length):
         if offset not in self.index:
             return None
-        
+
         length1, crc, st, _ = self.index[offset]
         data = os.pread(self.datafile, length, offset)
         if st == -1:
@@ -86,13 +121,13 @@ class _MachiGen:
 
         if length == length1:
             assert crc == binascii.crc32(data)
-            
+
         return data
-    
+
     def trim(self, offset, length):
         if offset not in self.index:
             return None
-        
+
         length1, crc, st, index_pos = self.index[offset]
         data = os.pread(self.datafile, length, offset)
         if st == -1:
@@ -109,26 +144,51 @@ class _MachiGen:
         self.index[offset] = (length, crc, -1, index_pos)
 
         self._ref -= 1
-    
+
     def close(self):
         os.close(self.indexfile)
         os.close(self.datafile)
-        os.remove(self.indexname)
-        os.remove(self.dataname)
+        if self.temp:
+            os.remove(self.indexname)
+            os.remove(self.dataname)
+
+
+def _parse_file(path):
+    print('Previous era data {} found. Loading...'.format(path))
+    dir, file = os.path.split(path)
+    base, ext = os.path.splitext(file)
+    assert ext == '.machi'
+    gen = int(base)
+    return gen, _MachiGen(dir, gen, creat=False, temp=False)
 
 
 class MachiStore:
-    def __init__(self, dir='/tmp', maxlen=1024*1024):
-        # TODO(kuenishi): Load existent files
-        self.gen = 0
+    def __init__(self, dir='/tmp', maxlen=1024*1024, temp=False):
         self.dir = dir
-        self.maxlen = maxlen
-        
-        self.front = _MachiGen(self.dir, self.gen)
-        assert self.front is not None
-        
+        self.temp = temp
+
         self.back = {}
-        
+
+        # TODO(kuenishi): Load existent files
+        if self.temp:
+            for file in glob.iglob(os.path.join(dir, '[0-9]*.machi')):
+                raise RuntimeError("temp but file exists: " + file)
+            for file in glob.iglob(os.path.join(dir, '[0-9]*.machd')):
+                raise RuntimeError("temp but file exists: " + file)
+
+        for file in glob.iglob(os.path.join(dir, '[0-9]*.machi')):
+            gen, machi_gen = _parse_file(file)
+            self.back[gen] = machi_gen
+
+        self.gen = 0
+        if self.back:
+            self.gen = max(self.back.keys()) + 1
+
+        self.maxlen = maxlen
+
+        self.front = _MachiGen(self.dir, self.gen, temp=self.temp)
+        assert self.front is not None
+
         self.rwlock = RWLock()
 
     def append(self, data: bytes): # -> File, offset, length
@@ -136,29 +196,38 @@ class MachiStore:
             result = self.front.append(data)
             gen = self.gen
             if len(self.front) >= self.maxlen:
-            
+
                 if self.front.ref == 0:
                     self.front.close()
                 else:
                     self.back[self.gen] = self.front
                     self.gen += 1
 
-                self.front = _MachiGen(self.dir, self.gen)
-        
+                self.front = _MachiGen(self.dir, self.gen,
+                                       temp=self.temp)
+
             offset, length = result
             return gen, offset, length
-    
+
+    def keys(self):
+        with self.rwlock.rdlock():
+            for key in self.front.keys():
+                yield key
+            for gen in self.back.keys():
+                for key in self.back[gen].keys():
+                    yield key
+
     def get(self, gen, offset, length) -> Optional[bytes]:
         with self.rwlock.rdlock():
             if gen == self.gen:
                 return self.front.get(offset, length)
-        
+
             f = self.back.get(gen)
-            # print(f)
+            # print(f, bool(f))
             # print(self.back)
             if f:
                 return f.get(offset, length)
-        
+
             return None
 
     def trim(self, gen, offset, length):
@@ -166,7 +235,7 @@ class MachiStore:
             if gen == self.gen:
                 self.front.trim(offset, length)
                 return
-        
+
             f = self.back.get(gen)
 
             if f:
@@ -174,7 +243,7 @@ class MachiStore:
                 if f.ref == 0:
                     f.close()
                     self.back.pop(gen)
-        
+
             return None
 
     def close(self):
